requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
crispy-bootstrap5==2024.2
Django==4.2.11
django-crispy-forms==2.1
django-filter==24.2
djangorestframework==3.15.1
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing-extensions==4.11.0
